autorebase: Specify the base ref name for rebase
Specifying the `base` argument for the `peter-evans/rebase` action
should make it consider only those PRs that can actually be affected by
the push (the ones that are targeted to the updated branch).  This stops
the rebase action from doing duplicate work, e.g., for its own pushes.

Note that the `pull_request_target` events also have `github.ref_name`
set to the base branch of the corresponding PR — in this case the
filtering will also work properly (the PR that caused the event should
still be found).