ci, flake: Major rewrite of matrix generation
The flake code is finally rewritten to split the `lib` part into
separate files, and some functions are factored out to be accessible
externally.

The CI matrix generation is redone to generate a list of `item` elements
for all build stages; those elements are now JSON objects potentially
containing multiple keys with list values, so that a single build job
may now build multiple items at once.  This feature would be useful to
group smaller packages together, because for many packages the time to
set up the build environment would be larger than the time to actually
build the package.

The CI job is still fake, but redone to accept the new data format
(currently just for the flake build part; the NUR build will be figured
out later).