ci: Grab build logs from /nix/var/log/nix/drvs
In some cases the build logs of the exposed package is not really
interesting, because it is just a wrapper for a real package which is
not exposed directly.  Parsing the output of `nix build --dry-run` is
not a very clean solution (the list of derivations to be build is
printed only on stderr, and the format is not intended to be machine
readable), so just grab all logs from `/nix/var/log/nix/drvs` (assuming
that the CI environment is used just for a single build operation and
therefore does not have unrelated build logs there).