ci: Import `flake-utils` directly for NUR
Unfortunately, using `edolstra/flake-compat` does not actually makes it
possible to load even some of the flakes listed in `inputs` without
actually downloading all inputs, so it would result in downloading and
importing an extra copy of Nixpkgs.  Import the `numtide/flake-utils`
flake directly instead (this is rather simple, because that flake itself
does not have any inputs).