Copy parts of `numtide/flake-utils` for NUR
Unfortunately, NUR forbids any form of external URL access during the
evaluation of repositories, therefore using functions from external
libraries like `numtide/flake-utils` to build the package list for NUR
is not possible.  For now just copy the implementation of some functions
(`flattenTree` and `filterPackages`) to a local file, so that they would
be available when evaluating the NUR repo.