ci: Use `nix-build-uncached`
Building all packages with `nix build` consumes excessive amount of time
and bandwidth when those packages did not really change and can be
downloaded from binary caches, because `nix build` will then download
all those packages.  Use the `nix-build-uncached` wrapper to detect that
and build only those packages that are not cached yet.

Unfortunately, `nix-build-uncached` always uses the raw `nix-build`
command, therefore it does not support flakes directly (although the
previous `nix build --dry-run --json` mitigates that) and does not
support the `--json` option to produce machine readable output.  Because
of that the feature to report the closure sizes of built packages is
removed (although that feature required all packages to be fetched
anyway, because the size can be determined only for the packages present
in the local Nix store).