flake, nur, ci: Export the whole nested package set to NUR
Don't try to apply `flattenTree` and `filterPackages` to the package set
which gets exported to NUR — just keep broken or unsupported packages
there, as they would be in Nixpkgs too.

The conversion to a flat set form is still needed to build the CI job
matrix though; this requires some extra library code.  Probably the
package filtering would need to be added there too; currently that part
is not yet implemented.