ci: Add yet another NUR build job to fix dependencies
Use the `nur1`->`nur3` jobs for `nixpkgs-unstable`->`nixos-unstable`,
and the `flake`->nur2` jobs for the NUR channel which is the same as the
main `nixpkgs` branch of the flake.  This way the `nixos-unstable` NUR
build does not wait for the end of the flake build if the flake uses one
of the stable branches, and the NUR build for the main channel does not
wait on the `nixpkgs-unstable` NUR build.