ci: Perform the NUR evaluation check in `setup`
The NUR evaluation check is important, because it can catch cases when
the evaluation process tries to access the network (which is not allowed
in NUR).  The `setup` job is probably a good enough place for it
(although it is single-threaded, apparently most of the slowness comes
from downloading the channel tarballs, and those tarballs would be
needed for preparing the matrix anyway, so the extra slowdown would be
minimal).