ci: Add support for running flake checks at setup time
Some flake checks (e.g., the `pre-commit` check which looks for
formatting violations) are so trivial that running them in a separate
job does not make much sense (the job setup will consume much more time
than the actual check).  Add support to run such checks in the setup job
instead (this also makes it easier to abort the CI workflow in case of
errors, which may be the desired behavior).

(At the moment the `flake.setupChecks` job is not generated by the flake
code, so the “Run setup checks” step should be skipped.)